requirements.txt
django==2.0.8
psycopg2
djangorestframework
django-jinja
django-babel
django-multiselectfield
django-user-language-middleware
ddt
seleniumbase